Added click commands, fixed methods

+++ App/models/coordinator.py
@@ -1,5 +1,5 @@
 from sqlalchemy.sql.expression import func
-from App.models import User
+from App.models import *
 from App.database import db
 
 class Coordinator(User):
@@ -16,27 +16,32 @@ class Coordinator(User):
 
 
     def add_competition(self, competition_name, start_time, end_time):
-        competition = Competition.query.filter_by(name = competition_name)
+        competition = Competition.query.filter_by(name = competition_name).first()
         if competition:
             return None
 
         new_competition = Competition(name = competition_name, organization_id = self.organization_id, start_time = start_time, end_time = end_time)
-        self.competitions.append(new_competition)
-        db.session.add(self)
+
+        organization = Organization.query.get(self.organization_id)
+        organization.competitions.append(new_competition)
+
+        db.session.add(new_competition)
         db.session.commit()
         return new_competition
 
 
     def edit_competition(self, competition_id, competition_name, start_time, end_time):
         competition = Competition.query.get(competition_id)
-        competition.name = competition_name
-        competition.start_time = start_time
-        competition.end_time = end_time
-        competition.organization_id = self.organization_id
-        
-        db.session.add(self)
-        db.session.commit()
-        return competition
+        if competition.organization_id == self.organization_id:
+            competition.name = competition_name
+            competition.start_time = start_time
+            competition.end_time = end_time
+            competition.organization_id = self.organization_id
+            
+            db.session.add(competition)
+            db.session.commit()
+            return competition
+        return None
 
 
     def remove_competition(self, competition_id):
@@ -44,27 +49,40 @@ class Coordinator(User):
         if not competition:
             return None
         
-        db.session.delete(competition)
-        db.session.commit()
-        return True
+        if competition.organization_id == self.organization_id:
+            db.session.delete(competition)
+            db.session.commit()
+            return True
+        return None
 
 
     def add_team(self, competition_id, team_name, points, time_taken):
-        #do i have to query the bridge table/relationship?
-        team = Team.query.filter_by(team_name = team_name)
+        team = Team.query.filter_by(team_name = team_name).first()
         if team:
             return None
-        
+
         new_team = Team(competition_id, team_name, points, time_taken)
         competition = Competition.query.get(competition_id)
         competition.teams.append(new_team)
-        db.session.add(self)
+        db.session.add(new_team)
         db.session.commit()
         return new_team
 
 
-    #def edit_team(self, competition_id, team_id, team_name, points, time_taken):
-    
+    def edit_team(self, competition_id, team_id, team_name, points, time_taken):
+        
+        team = Team.query.get(team_id)
+        if team:
+            team.competition_id = competition_id
+            team.team_name = team_name
+            team.points = points
+            team.time_taken = time_taken
+
+            db.session.add(team)
+            db.session.commit()
+            return team
+        return None
+
 
     def remove_team(self, team_id):
         team = Team.query.get(team_id)
@@ -96,5 +114,5 @@ class Coordinator(User):
             'username': self.username
         }
 
-    def repr():
+    def repr(self):
         return f'<Coordinator {self.coordinator_id} : {self.first_name} {self.last_name} {self.username} - {self.email}>'
